==> scripts/4. Get Weather Data - Hourly Weather .ipynb <==
# Read Libs

import pandas as pd
import datetime as dt
import numpy as np
import requests as req
import os
from bs4 import BeautifulSoup
import html5lib
import time
import re

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.chrome.options import Options

# display and output settings
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

from IPython.display import display, HTML




display(HTML('''
<style>
    .container { width:85% !important; }
</style>
'''))

pd.set_option('display.min_rows', 100)

# Read in Tournament & Weather URL Details

path_tournament_detail = "../data/transformed_data/tournament_info/2024-10-06 16h20m45s/tournament_details_transformed.csv"
path_tournament_scores = "../data/transformed_data/tournament_scores/2024-10-06 16h20m45s/tournament_scores_transformed.csv"
path_weather_urls = "../data/raw_data/weather/weather_city_base_urls/2024-10-26 12h17m18s/weather_city_base_urls.csv"

df_tournament_details = pd.read_csv(path_tournament_detail)
df_tournament_scores = pd.read_csv(path_tournament_scores)
df_weather_urls = pd.read_csv(path_weather_urls)

df_tournament_details.head()
df_weather_urls.head()
df_tournament_scores.head()

# Clean Weather Urls

df_weather_urls['base_url'].apply(lambda x: len(str(x))).value_counts()


# replace urls if they are not valid by rules of length being less than 62
# takes away urls that could not find a city and are null
df_weather_urls['base_url'] = df_weather_urls['base_url'].apply(lambda x: None if len(str(x)) < 62 else x)

# Create tournament calendar | Round by Round Dates

df_tournament_details.head(3)
df_tournament_details.info()

def create_array_of_dates(start_date, end_date):
    """
    creates an array of dates between start_date and end_date
    """
    try:

        start_date = dt.datetime.strptime(start_date, "%Y-%m-%d")
        end_date = dt.datetime.strptime(end_date, "%Y-%m-%d")
        end_date = end_date + dt.timedelta(days=1)

        date_generated = [start_date + dt.timedelta(days=x) for x in range(0, (end_date-start_date).days)]
        date_generated = [date.strftime("%Y-%m-%d") for date in date_generated]
        return date_generated
    except:
        return None

df_tournament_details['round_dates']=\
df_tournament_details.apply(lambda x: create_array_of_dates(x['start_date'], x['end_date']), axis=1)

# explode the dates to have one row per date
df_tournament_details = df_tournament_details.explode('round_dates')

# create a column for the round of the tournament
df_tournament_details['round'] = df_tournament_details.groupby('event_id').cumcount() + 1

## Merge dataframes

df_merged_tournament_weather = pd.merge(df_tournament_details, df_weather_urls, left_on="city", right_on="City", how='left')

def get_url_weather_date(base_url, date):
    """
    creates a url for the weather for a specific date
    """
    try:
        match_date = re.findall("\d{4}-\d{1,2}-\d{1,2}", base_url)[0]

        return base_url.replace(match_date, date)
    except:
        return None    
    

df_merged_tournament_weather.head(3)

df_merged_tournament_weather['weather_url'] = df_merged_tournament_weather\
    .apply(lambda x: get_url_weather_date(x['base_url'], x['round_dates']), axis=1)


# Get Table from URLs

### Create Batches for Dataframe

# filter out the non null weather urls and save the dataframe 
df_merged_tournament_weather_non_null_url =  \
df_merged_tournament_weather[~df_merged_tournament_weather['weather_url'].isnull()].reset_index(drop=True)

# define rows per batch
rows_per_batch = 50

# Define the number of batches
num_batches = len(df_merged_tournament_weather_non_null_url) / rows_per_batch
num_batches = int(np.ceil(num_batches))

# Split the dataframe into batches
batches = np.array_split(df_merged_tournament_weather_non_null_url, num_batches)

print(f'\n ---------- \n number of batches are: {num_batches}')
for batch_number, batch_df in enumerate(batches):
    print(batch_number, len(batch_df))
    

### Launch Web Crawler

# selenium chromedriver
# Specify the path to the ChromeDriver executable
chrome_driver_path = '../drivers/chromedriver_arm'

#link for solution to cannot open chromedriver
# https://stackoverflow.com/questions/60362018/macos-catalinav-10-15-3-error-chromedriver-cannot-be-opened-because-the-de
# xattr -d com.apple.quarantine chromedriver

# Set the options for the webdriver to avoid rendering ads
# chrome_options = Options()
# chrome_options.add_argument('--headless')
# chrome_options.add_argument('--disable-gpu')

#open the browser
# driver = webdriver.Chrome(executable_path=chrome_driver_path, options=chrome_options)
driver = webdriver.Chrome(executable_path=chrome_driver_path)

# import warnings filter to suppress warnings
import warnings
warnings.filterwarnings('ignore')

# import to clear output after each batch to understand where the script is
from IPython.display import clear_output

for batch_number, batch_df in enumerate(batches):

    print(f"Batch Number: {batch_number} \n -------------------")

    # create a timestamp for the batch
    timestamp = dt.datetime.now().strftime("%Y-%m-%d %Hh%Mm%Ss")
    path_weather_hourly_raw_data = "../data/raw_data/weather/hourly_weather/"
    
    # create a directory for the batch
    os.makedirs(f"{path_weather_hourly_raw_data}/{batch_number}/{timestamp}")

    # initalize a counter for the batch and a list to store the hourly weather data for the batch
    counter = 0
    df_list_hourly_weather_data = []

    # loop through batch df valid weather urls
    for index, row in batch_df.iterrows():
    
        # for each row in dataframe that has a weather url get the event id, round and weather url
        weather_url = row['weather_url']
        event_id = row['event_id']
        round = row['round']
        tournament_city = row['city']
        counter += 1

        print(f"{counter} | Event ID: {event_id}, Round: {round}, Weather URL: {weather_url}")

        ### Use selenium to load the page and extract the hourly weather data

        # URL of the page with dynamically loaded tables
        url = weather_url

        #set the page load timeout to 20 seconds
        driver.set_page_load_timeout(15)
        
        # Load the page
        #try to load the page if it fails wait 5 seconds and try again
        try:
            driver.get(url)
        except:
            print('retrying to load page')
            time.sleep(5)
            driver.get(url)

        # Xpath to the table with the hourly weather data
        x_path_observation = "/html/body/app-root/app-history/one-column-layout/wu-header/sidenav/mat-sidenav-container/mat-sidenav-content/div[2]/section/div[2]/div[1]/div[5]/div[1]/div/lib-city-history-observation/div/div[2]/table"

        try:
            WebDriverWait(driver, 15).until(
                EC.presence_of_element_located((By.XPATH, x_path_observation))
            )
            
            # Extract the page source after javascript has loaded
            html = driver.page_source
            soup = BeautifulSoup(html, 'html.parser')

            # Find table within the div with the class 'observation-table'
            table_observation = soup.find('div', class_='observation-table').find('table')

            # Find the actual city name from the page
            weather_history_span = soup.find('span', string=re.compile("Weather History"))
            actual_city = weather_history_span.text.replace("Weather History", "").strip()
        

            # Convert the hourly weather data html table to a pandas DataFrame
            df_hourly_weather = pd.read_html(str(table_observation))[0]

            #clean up rows that have nulls in the Time column
            df_hourly_weather.dropna(subset=['Time'], inplace=True)

            # adding key columns to map back to the tournament details dataframe
            df_hourly_weather['event_id'] = event_id
            df_hourly_weather['round'] = round
            df_hourly_weather['weather_url'] = weather_url

            #adding city columns to help compare later if that city pulled for the weather is near the tournament city
            df_hourly_weather['tournament_city'] = tournament_city
            df_hourly_weather['actual_weather_city'] = actual_city

            # Append the hourly weather data to the list of dataframes
            df_list_hourly_weather_data.append(df_hourly_weather)

        except:
            print("page did not load in time for see hourly weather table")

    # Concatenate the list of hourly weather dataframes into a single dataframe
    df_hourly_weather_data = pd.concat(df_list_hourly_weather_data)

    # save the hourly weather data to a csv file
    df_hourly_weather_data.to_csv(f"{path_weather_hourly_raw_data}/{batch_number}/{timestamp}/hourly_weather_data.csv", index=False)

    print(f"batch: {batch_number} Saved to file")
    # clear the output to show the next batch
    clear_output()
