==> scripts/7. Final Transformations - Store Final Data.ipynb <==
# Read Libs

import pandas as pd
import datetime as dt
import numpy as np
import os
import re
import shutil

# display and output settings
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

from IPython.display import display, HTML


display(HTML('''
<style>
    .container { width:85% !important; }
</style>
'''))

pd.set_option('display.min_rows', 100)

# Read Files In

path_transform_data_tournament_info= "../data/transformed_data/tournament_info"
path_transform_data_tournament_scores= "../data/transformed_data/tournament_scores"
path_transform_data_tournament_weather= "../data/transformed_data/weather_data/clean_weather_values"


list_paths_data = [path_transform_data_tournament_info, 
                   path_transform_data_tournament_scores,
                   path_transform_data_tournament_weather]

# getting latest file in each directory

list_file_path_data = []
for directory in list_paths_data:

    batch_runs = list(filter(lambda x: ".DS_Store" not in x, os.listdir(directory)))
    
    latest_run = max(batch_runs)

    file_path = os.listdir(f"{directory}/{latest_run}")

    latest_file = list(filter(lambda x: ".DS_Store" not in x, file_path))[0]

    latest_file_path = f"{directory}/{latest_run}/{latest_file}"

    list_file_path_data.append(latest_file_path)


details = pd.read_csv(list_file_path_data[0])
scores = pd.read_csv(list_file_path_data[1])
weather = pd.read_csv(list_file_path_data[2])

## Preview

details.head()
scores.head()
weather.head()

# Transformations

### Create Round Dates from Details | Taking Logic from Notebook 4

def create_array_of_dates(start_date, end_date):
    """
    creates an array of dates between start_date and end_date
    """
    try:

        start_date = dt.datetime.strptime(start_date, "%Y-%m-%d")
        end_date = dt.datetime.strptime(end_date, "%Y-%m-%d")
        end_date = end_date + dt.timedelta(days=1)

        date_generated = [start_date + dt.timedelta(days=x) for x in range(0, (end_date-start_date).days)]
        date_generated = [date.strftime("%Y-%m-%d") for date in date_generated]
        return date_generated
    except:
        return None


dates_by_round_details = details.copy()

dates_by_round_details['round_dates']=\
dates_by_round_details.apply(lambda x: create_array_of_dates(x['start_date'], x['end_date']), axis=1)

# explode the dates to have one row per date
dates_by_round_details = dates_by_round_details.explode('round_dates')

# create a column for the round of the tournament
dates_by_round_details['round'] = dates_by_round_details.groupby('event_id').cumcount() + 1

dates_by_round_details.head()

#### Create Key for Joining on Round and Event

dates_by_round_details['Key_Event_Rd'] = "E"+ dates_by_round_details['event_id'].astype(str) + "R" + dates_by_round_details['round'].astype(str)

#### Clean up Columns Needed

cols_to_keep = ['event_id', 'round', 'Key_Event_Rd', 'round_dates']
dates_by_round_details[['event_id', 'round', 'Key_Event_Rd', 'round_dates']].head()

dates_by_round_details = dates_by_round_details[cols_to_keep]


### Melt Scores to Have Event, Round #, and Score

scores_by_round = scores.melt(
    id_vars=['event_id', 'PLAYER'],
    value_vars=[f'R{i}' for i in range(1, 5)],
    var_name='round',
    value_name='round_score'
)
# join par to scores by round

scores_by_round = scores_by_round\
.merge(details[['event_id', 'par']], on='event_id', how='left')

scores_by_round['round'] = scores_by_round['round'].str.replace('R', '').astype(int)



scores_by_round['round'].value_counts()
scores_by_round.head()


def to_par(round_score, par):
    try:
        return int(round_score) - int(par)
    except:
        return np.nan

scores_by_round['to_par'] = scores_by_round\
    .apply(lambda x: to_par(x['round_score'], x['par']), axis=1)


#### Adding Key


scores_by_round['Key_Event_Rd'] = "E"+ scores_by_round['event_id'].astype(str) + "R" + scores_by_round['round'].astype(str)

scores_by_round.head()

### Adding Key to Weather Data

weather["Key_Event_Rd"] = "E"+ weather['event_id'].astype(str) + "R" + weather['round'].astype(str)

### Create Major Tournament Flag on Details

from fuzzywuzzy import fuzz, process

def create_major_flag(name):
    """
    returns 1 if the tournament is a major, 0 otherwise
    """

    major_tournaments = ["Masters", "PGA Championship", "The Open", "The Open Championship", "U.S. Open Championship", "U.S. Open"]
    major_tournaments = [x.lower() for x in major_tournaments]

    match = process.extractOne(query= name.lower(), 
                   choices = major_tournaments,
                   score_cutoff= 90)
    
    # manual list to not count some tournaments as majors to aid fuzzy matches
    list_do_not_count = [
    'Senior PGA Championship',
    'Sportsbet Australian Masters',
    'Commercialbank Qatar Masters',
    'BMW PGA Championship',
    'SAS Masters',
    'Omega European Masters',
    'Madrid Masters',
    'Portugal Masters',
    'CASTELLO MASTERS Costa Azahar',
    'JBWere Masters',
    'Commercial Bank Qatar Masters presented by Dolphin',
    'Avantha Masters',
    'BMW PGA CHAMPIONSHIP',
    'Nordea Scandinavian Masters',
    'Andalucia Masters',
    'Commercialbank Qatar Masters presented by Dolphin ',
    'Nordea Scandanavian Masters',
    'Bankia Madrid Masters',
    'Commercialbank Qatar Masters presented by Doplhin ',
    'Nordea Masters',
    'BMW Masters',
    'Commercial Bank Qatar Masters',
    'Wegmans LPGA Championship',
    'D+D REAL Czech Masters ',
    'AfrAsia Bank Mauritius Open',
    'D+D Real Czech Masters',
    'British Masters',
    "KPMG Women's PGA Championship",
    'Australian PGA Championship',
    'British Masters supported by Sky Sports',
    'KitchenAid Senior PGA Championship',
    'British Masters Supported by Sky Sports',
    'Andalucia Valderrama Masters hosted by the Sergio ',
    'DAP Championship',
    'SAS Championship',
    'HSBC New Zealand PGA Championship',
    'Jacob\'s Creek Open Championship',
    'U.S. Senior Open Championship',
    'SA Open Championship',
    'The Senior Open Championship'
    ]
    


    if match:
        if name not in list_do_not_count:
            return 1
    else:
        return 0

details['is_major'] = details['tournament_title'].apply(create_major_flag)



def normalize_major_name(major_flag, name):

    if major_flag == 1:
        if "masters" in name.lower():
            return "The Masters"
        elif "pga championship" in name.lower():
            return "PGA Championship"
        elif "u.s. open" in name.lower():
            return "U.S. Open"
        elif "the open" in name.lower():
            return "The Open Championship"
    else:
        return name   

details['tournament_title'] = details.apply(lambda x: normalize_major_name(x['is_major'], x['tournament_title']), axis=1)  
details[details['is_major'] == 1]['tournament_title'].value_counts()       

#### Strip whitespaces in cities

details['city'] = details['city'].str.strip()


#### Add lat/long for Viz Software 

from geopy.geocoders import Nominatim
from geopy.distance import geodesic

def geocode_city(city_name):
    """
    Geocode a city name to obtain its geographical location.
    This function uses the Nominatim geocoding service to convert a city name into its corresponding geographical location (latitude and longitude).
    Args:
        city_name (str): The name of the city to geocode.
    Returns:
        location (Location or None): A Location object containing the geographical coordinates of the city if successful, or None if the geocoding times out.
    Raises:
        GeocoderTimedOut: If the geocoding request times out.
    """

    geolocator = Nominatim(user_agent="my_geocoder")
    try:
        location = geolocator.geocode(city_name, timeout=10)
        return location
    except GeocoderTimedOut:
        print(f"Geocoding timed out for city: {city_name}")
        return None

# compressing dataframe to just cities to get 400 rows instead of 2000
city_agg = details[['city', 'event_id']]\
    .groupby('city').count().reset_index()

city_agg['latitude'] = city_agg['city'].apply(geocode_city).apply(lambda x: x.latitude if x else None)
city_agg['longitude'] = city_agg['city'].apply(geocode_city).apply(lambda x: x.longitude if x else None)

details = details\
        .merge(city_agg[['city', 'latitude', 'longitude']], 
           on='city', 
           how='left')

# Preview All Data Tables

# tournament detail will have many rounds and dates
# each round will have many players and scores
# each round will have hourly weather details

details.head()
dates_by_round_details.head()
scores_by_round.head()
weather.head()

# Store to Final Folder


final_data_folder = "../data/final_data"

# write to final folder
details.to_csv(f"{final_data_folder}/tournament_info.csv", index=False)

dates_by_round_details.to_csv(f"{final_data_folder}/dates_event_round_dates.csv", index=False)

scores_by_round.to_csv(f"{final_data_folder}/player_scores_by_round.csv", index=False)

scores.to_csv(f"{final_data_folder}/scores_summary_final.csv", index=False)

weather.to_csv(f"{final_data_folder}/weather_data.csv", index=False)
